Testing Authentication

+++ RSMSessionsEnterpriseIntegrations/Application/DTOs/SalesOrderHeader/CreateSalesOrderHeaderDto.cs
@@ -2,11 +2,11 @@
 {
     public class CreateSalesOrderHeaderDto
     {
-        public short RevisionNumber { get; set; }
+        public byte RevisionNumber { get; set; }
         public DateTime OrderDate { get; set; }
         public DateTime DueDate { get; set; }
         public DateTime? ShipDate { get; set; }
-        public short Status { get; set; }
+        public byte Status { get; set; }
         public bool OnlineOrderFlag { get; set; }
         public string? PurchaseOrderNumber { get; set; }
         public string? AccountNumber { get; set; }
