+++ RSMSessionsEnterpriseIntegrations/Application/DTOs/SalesOrderHeader/GetSalesOrderHeaderDto.cs
@@ -3,11 +3,11 @@
     public class GetSalesOrderHeaderDto
     {
         public int Id { get; set; }
-        public short? RevisionNumber { get; set; }
+        public byte? RevisionNumber { get; set; }
         public DateTime? OrderDate { get; set; }
         public DateTime? DueDate { get; set; }
         public DateTime? ShipDate { get; set; }
-        public short? Status { get; set; }
+        public byte? Status { get; set; }
         public bool? OnlineOrderFlag { get; set; }
         public string? SalesOrderNumber { get; }
         public string? PurchaseOrderNumber { get; set; }
