+++ RSMSessionsEnterpriseIntegrations/Domain/Models/SalesOrderHeader.cs
@@ -4,11 +4,11 @@
 
     public class SalesOrderHeader : Entity
     {
-        public short? RevisionNumber { get; set; }
+        public byte? RevisionNumber { get; set; }
         public DateTime? OrderDate { get; set; }
         public DateTime? DueDate { get; set; }
         public DateTime? ShipDate { get; set; }
-        public short? Status { get; set; }
+        public byte? Status { get; set; }
         public bool? OnlineOrderFlag { get; set; }
         public string? SalesOrderNumber { get; }
         public string? PurchaseOrderNumber { get; set; }
